+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - A major rewrite in formula view to utilize select2 (WIP)
 - Various improvements
 - Removed sorting action from properties and actions columns in formula
+- Make sure the dilutant field isn't enabled when purity is at 100%
 
 ### Version 4.5
 - Fix core upgrade